tox_count_friendlist()
toxcore eac0d435f35db468435fa596a3b00b593d6bf3e9

>tox_self_get_friend_list_size()
toxcore fa7c21f95c78eabf3c0f0d247cd7f28ec47eff3d

+++ src/friendlist.h
@@ -86,7 +86,7 @@ typedef struct {
 typedef struct {
     int num_selected;
     int max_idx;    /* 1 + the index of the last friend in list */
-    int num_friends;
+    size_t num_friends;
     int *index;
     ToxicFriend *list;
 } FriendsList;
